lots of pre-ump file reorganization
