+++ static/js/src/connected-elements/Hidden.vue
