+++ static/js/src/connected-elements/ManualPay.vue
