+++ static/js/src/connected-elements/ManualSubmit.vue
