+++ static/js/src/entry/donate/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 import App from '../../App.vue';
-import FormModule from '../../store/modules/form';
+import connectedFormModule from '../../store/modules/connected-form';
 import { createRouter, bindRouterEvents } from './router';
 import cssClasses from '../../mixins/global/cssClasses';
 import gtm from '../../mixins/global/gtm';
@@ -15,7 +15,7 @@ Vue.mixin(gtm);
 
 const store = new Vuex.Store({
   modules: {
-    baseForm: FormModule,
+    baseForm: connectedFormModule,
   },
 });
 const router = createRouter();
